Release a reference implementation of PA1
Identify necessary APIs for PA1

This commit prints the phases of a fanin in all cubes.
It uses the APIs necessary for PA1.
The circuit full_adder.blif is added as a simple test.

In essence, students need to know:
1) How SOP is stored as a char*
2) How to decide whether the SOP represents onset or offset
3) How to read the phase of a variable in a cube

Modify the full adder blif file

Rename the PI/PO (c-in, c-out); delete redundant cube "111 1" from MAJ

Fix null pointer exception when the network does not have SOP

Delete full_adder.blif in the root directory of ABC

Define a testing alias `lsv' in abc.rc

Finish a reference implemenation for PA1