Handle the output format
As an extracted cone from the original circuit might have different PI orders,
the mapping has to be recorded.
For the PIs that do not appear in the TFI cone of a PO,
the PO is both positive and negative unate with respect to them.
To debug, a flag to toggle the output formats is added.
Using `lsv_print_pounate -l` outputs the same format as command `print_unate -v`.